+++ src/hooks/LevelInfoLayer.cpp
@@ -14,8 +14,6 @@ class $modify(MyLevelInfoLayer, LevelInfoLayer) {
         CCLabelBMFont* m_label;
 
         CCPoint const m_position = SettingsManager::Display.getPosition();
-
-        async::TaskHolder<utils::web::WebResponse> m_listener;
     };
 
     bool init(GJGameLevel* level, bool challenge) {
@@ -191,27 +189,59 @@ class $modify(MyLevelInfoLayer, LevelInfoLayer) {
                 } else {
                     static std::string_view constexpr placeholder = "Sent: Loading...";
                     labelContent << placeholder << std::endl;
-                    
-                    auto req = utils::web::WebRequest()
-                        .userAgent(std::string(Utils::GetUserAgent()));
 
                     auto levelID = static_cast<int>(level->m_levelID);
+                    
+                    async::spawn(
+                        [levelID]() -> arc::Future<Result<bool, void>> {
+                            // First try requesting to my own cache API specifically for
+                            // the sent state
+                            auto req = co_await utils::web::WebRequest()
+                                .userAgent(Utils::GetUserAgent())
+                                .timeout(std::chrono::seconds(3))
+                                .get(fmt::format("https://sdbc.m336.dev/level/{}", levelID));
+                            auto body = req.json().unwrapOrDefault();
+                            auto error = body["error"].asString().unwrapOrDefault();
 
-                    // This will call SendDB's API to know if the level was sent or
-                    // not. The actual sent field will be updated after the request
-                    // is completed
-                    m_fields->m_listener.spawn(
-                        req.get(fmt::format("https://api.senddb.dev/api/v1/level/{}", levelID)),
-                        [self, levelID](utils::web::WebResponse result) {
+                            // If that doesn't work, fallback to the original SendDB API
+                            if (!req.ok() || body.size() <= 0 || error.size() > 0) {
+                                log::warn(
+                                    "Failed requesting to the cache API ({}), fallback to the original SendDB API",
+                                    error.size() > 0 ? error : req.errorMessage()
+                                );
+
+                                req = co_await utils::web::WebRequest()
+                                    .userAgent(Utils::GetUserAgent())
+                                    .timeout(std::chrono::seconds(3))
+                                    .get(fmt::format("https://api.senddb.dev/api/v1/level/{}", levelID));
+                                body = req.json().unwrapOrDefault();
+                                error = req.errorMessage();
+
+                                // If that still doesn't work, don't go further
+                                if (!req.ok() || body.size() <= 0) {
+                                    log::error(
+                                        "Failed requesting to the SendDB API: {}",
+                                        error.size() > 0 ? error : req.string().unwrap()
+                                    );
+                                    co_return Err();
+                                }
+
+                                // For the SendDB API, just check if the sends object is more than 0
+                                co_return Ok(body["sends"].size() > 0 ? true : false);
+                            }
+
+                            // For the cache, it's directly indicated in the sent boolean
+                            co_return Ok(body["sent"].asBool().unwrap());
+                        },
+                        [self, levelID](Result<bool, void> result) {
                             if (!self->m_fields->m_label)
                                 return;
                             
-                            matjson::Value body = result.json().unwrapOrDefault();
+                            bool failed = !result.isOk();
+                            bool sent = failed ? false : result.unwrap();
 
-                            bool isSent = body.size() > 0 &&
-                                body.contains("sends") && body["sends"].size() > 0;
-                            if (SettingsManager::Other.enableSentCache && body.size() > 0)
-                                SentCacheManager::SaveLevel(levelID, isSent);
+                            if (SettingsManager::Other.enableSentCache && !failed)
+                                SentCacheManager::SaveLevel(levelID, sent);
 
                             std::string labelContent = self->m_fields->m_label->getString();
                             size_t pos = labelContent.find(placeholder);
@@ -221,8 +251,8 @@ class $modify(MyLevelInfoLayer, LevelInfoLayer) {
                                     placeholder.length(),
                                     fmt::format(
                                         "Sent: {}",
-                                        body.size() > 0
-                                            ? isSent
+                                        !failed
+                                            ? sent
                                                 ? "Yes"
                                                 : "No"
                                             : "Failed"